# [android] remove Anticipator.cs for now