@kingces95 2020-01-07
You've removed the call sites so simplest thing would be to simply undelete the file.